устранение замечаний.

+++ core/src/main/java/ru/stepup/spring/coins/core/controllers/CoinsController.java
@@ -20,7 +20,7 @@ public class CoinsController {
 
     @PostMapping("/execute")
     public ExecuteCoinsResponse execute(@RequestBody ExecuteCoinsRequest request,
-                                        @RequestHeader("USERID") Integer userId) {
+                                        @RequestHeader("USERID") Long userId) {
         return coinsService.execute(request, userId);
     }
 }
