+++ core/src/main/java/ru/stepup/spring/coins/core/integrations/ProductIntegration.java
@@ -4,5 +4,5 @@ import ru.stepup.spring.coins.core.api.ProductResponse;
 
 
 public interface ProductIntegration {
-    ProductResponse getUserProduct(Integer userId);
+    ProductResponse getUserProduct(Long userId);
 }
